LayoutTests/platform/mac-wk1/fast/scrolling/scroll-animator-basic-events-expected.txt

CONSOLE MESSAGE: line 27: RenderLayer: mouseEnteredContentArea
CONSOLE MESSAGE: line 27: RenderLayer: mouseMovedInContentArea
CONSOLE MESSAGE: line 28: RenderLayer: mouseEnteredVerticalScrollbar
CONSOLE MESSAGE: line 28: RenderLayer: mouseMovedInContentArea
CONSOLE MESSAGE: line 29: RenderLayer: mouseIsDownInVerticalScrollbar
CONSOLE MESSAGE: line 30: RenderLayer: mouseIsUpInVerticalScrollbar
CONSOLE MESSAGE: line 30: RenderLayer: mouseMovedInContentArea
CONSOLE MESSAGE: line 31: RenderLayer: mouseExitedVerticalScrollbar
CONSOLE MESSAGE: line 31: RenderLayer: mouseExitedContentArea
PASS successfullyParsed is true

TEST COMPLETE
This is a scrollable div.
